+++ demo/src/app/components/+datepicker/default-picker-value/datepicker-default-picker-value.component.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'app-demo-datepicker-default-picker-value',
+    templateUrl: './datepicker-default-picker-value.component.html'
+})
+export class DemoDatePickerDefaultPickerValueComponent {
+    today = new Date();
+
+    defaultPickerValue = new Date('2022-03-03');
+
+    defaultRangePickerValue = [new Date('2022-03-03'), new Date('2022-04-04')];
+}