+++ src/dialog/dialog.module.ts
@@ -9,11 +9,13 @@ import { DialogBodyComponent } from './body/dialog-body.component';
 import { DialogFooterComponent } from './footer/dialog-footer.component';
 import { ThyButtonModule } from '../button';
 
-import { THY_DIALOG_DEFAULT_OPTIONS_PROVIDER } from './dialog.config';
+import { THY_DIALOG_DEFAULT_OPTIONS_PROVIDER, THY_DIALOG_LAYOUT_CONFIG_PROVIDER } from './dialog.config';
 import { ThyConfirmComponent } from './confirm/confirm.component';
 import { THY_CONFIRM_DEFAULT_OPTIONS_PROVIDER } from './confirm.config';
 import { ThyIconModule } from '../icon/icon.module';
 import { ThySharedModule } from '../shared';
+import { ThyFormModule } from '../form';
+import { FormsModule } from '@angular/forms';
 
 @NgModule({
     declarations: [
@@ -23,8 +25,22 @@ import { ThySharedModule } from '../shared';
         DialogFooterComponent,
         ThyConfirmComponent
     ],
-    imports: [CommonModule, ThySharedModule, PortalModule, OverlayModule, ThyButtonModule, ThyIconModule],
-    providers: [ThyDialog, THY_DIALOG_DEFAULT_OPTIONS_PROVIDER, THY_CONFIRM_DEFAULT_OPTIONS_PROVIDER],
+    imports: [
+        CommonModule,
+        ThySharedModule,
+        PortalModule,
+        OverlayModule,
+        ThyButtonModule,
+        ThyIconModule,
+        ThyFormModule,
+        FormsModule
+    ],
+    providers: [
+        ThyDialog,
+        THY_DIALOG_DEFAULT_OPTIONS_PROVIDER,
+        THY_CONFIRM_DEFAULT_OPTIONS_PROVIDER,
+        THY_DIALOG_LAYOUT_CONFIG_PROVIDER
+    ],
     entryComponents: [ThyDialogContainerComponent, ThyConfirmComponent],
     exports: [ThyDialogContainerComponent, DialogHeaderComponent, DialogBodyComponent, DialogFooterComponent]
 })