+++ src/directive/index.ts
@@ -1,9 +1,11 @@
 export * from './module';
 export * from './thy-autofocus.directive';
-export * from './thy-clo.directive';
+
+export * from './raster/thy-row.directive';
+export * from './raster/thy-col.directive';
+
 export * from './thy-ctrl-enter.directive';
 export * from './thy-enter.directive';
-export * from './thy-row.directive';
 export * from './thy-show';
 export * from './thy-stop-propagation.directive';
 export * from './thy-scroll.directive';